--- client/src/components/Article/index.js
@@ -1,50 +1,0 @@
-import React, { useEffect } from "react";
-import { withRouter } from "react-router-dom";
-import api from "../../lib/api";
-import styled from "styled-components";
-import ArticleSlider from "./ArticleSlider";
-import ArticleProfile from "./ArticleProfile";
-import ArticleDescription from "./ArticleDescription";
-import ArticleFooter from "./ArticleFooter";
-import { useState } from "react";
-
-const Article = ({ isModal, setModal, match }) => {
-  const [sliderData, setSliderData] = useState([]);
-  const [detail, setDetail] = useState([]);
-
-  let id = match.params.articles_id;
-  useEffect(() => {
-    window.localStorage.setItem("productId", id);
-    fetch("http://localhost:3000/data/sliderData.json")
-      .then((res) => res.json())
-      .then((res) => {
-        setSliderData(res.sliderData1);
-      });
-    // fetch(`${api}/product/${id}`)
-    //   .then((res) => res.json())
-    //   .then((res) => {
-    //     setDetail(res);
-    //   });
-    api.get(`/product/${id}`).then((res) => setDetail(res.data));
-  }, []);
-
-  return (
-    <ArticleWrapper>
-      <ArticleSlider
-        sliderData={sliderData}
-        isModal={isModal}
-        setModal={setModal}
-      />
-      <ArticleProfile />
-      <ArticleDescription detail={detail} />
-      <ArticleFooter detail={detail} id={id} />
-    </ArticleWrapper>
-  );
-};
-
-export default withRouter(Article);
-
-const ArticleWrapper = styled.div`
-  margin-top: 120px;
-  padding-bottom: 50px;
-`;
