+++ client/src/login/index.js
@@ -1,0 +1,13 @@
+import React from "react";
+import Layout from "../common/components/Layout";
+import Container from "./container";
+
+const Login = () => {
+  return (
+    <Layout>
+      <Container />
+    </Layout>
+  );
+};
+
+export default Login;
