+++ client/src/productDetail/components/Profile.js
@@ -2,7 +2,7 @@ import React from "react";
 import styled from "styled-components";
 import { Link } from "react-router-dom";
 
-const ArticleProfile = () => {
+const Profile = () => {
   return (
     <StyledArticleProfile>
       <div to="/#" className="article-profile-link">
@@ -37,7 +37,7 @@ const ArticleProfile = () => {
   );
 };
 
-export default ArticleProfile;
+export default Profile;
 
 const StyledArticleProfile = styled.section`
   width: 677px;
