+++ client/src/productDetail/index.js
@@ -1,0 +1,8 @@
+import React from "react";
+import Container from "./container";
+
+const productDetail = () => {
+  return <Container />;
+};
+
+export default productDetail;
