+++ client/src/redux/reducers/loginRedux.js
@@ -1,10 +1,9 @@
-
-import  api  from "../lib/api";
+import api from "../../common/api";
 const LOGIN = "login/LOGIN";
 const LOGOUT = "login/LOGOUT";
 const USERCHECK = "login/USERCHECK";
 
-export const loginMember = (id,pw) => ({ type: LOGIN , id: id, pw:pw });
+export const loginMember = (id, pw) => ({ type: LOGIN, id: id, pw: pw });
 export const decrement = () => ({ type: LOGOUT });
 export const usercheck = () => ({ type: USERCHECK });
 
@@ -12,7 +11,7 @@ const initialState = {
   pk: 0,
   user_id: "",
   name: "",
-  nick_name:"",
+  nick_name: "",
   tel: "",
   add: [],
   temper: 36.5,
@@ -35,14 +34,14 @@ const login = (state = initialState, action) => {
         if (response.status === 200) {
           alert("정상 로그인 되었습니다");
           user = await response.json();
-          return {...user};
+          return { ...user };
         } else {
           alert("응, 틀렸어~");
-          return {...state};
+          return { ...state };
         }
       });
     case LOGOUT:
-      return {...state};
+      return { ...state };
     default:
       return state;
   }
