+++ client/src/register/index.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Register = () => {
+  return <div></div>;
+};
+
+export default Register;
